Fix issues found in Copilot review
- `SignalHandlers.cpp`: swap inverted log messages for `crashing` variable
- `gtest_functions_stress.cpp`: fix UB in `reportProblem` bitmask, use
  `Problem` enum value instead of `function_idx` which can exceed 63
- `ColumnNullable.cpp`: fix unsigned underflow when `first == 0` by
  casting to `ssize_t` before subtracting
- `reverseUTF8.cpp`: validate UTF-8 sequence lengths before `memcpy`
  to prevent out-of-bounds writes on malformed input
- `trim.cpp`: use `ILLEGAL_COLUMN` instead of `LOGICAL_ERROR` for
  unexpected column type in user-facing validation
- `StorageGenerateRandom.cpp`: fix buffer overwrite in
  `fillBufferWithRandomPrintableASCIIBytes` when size is not a
  multiple of 4; fix UB when shift amount equals 64 in string
  length generation
- `readWkb.cpp`: include function name, expected type, and variant
  index in geometry type mismatch error message

+++ src/Columns/ColumnNullable.cpp
@@ -638,7 +638,7 @@ void ColumnNullable::updatePermutationImpl(IColumn::PermutationSortDirection dir
 
             ssize_t read_idx = last - 1;
             ssize_t write_idx = last - 1;
-            ssize_t begin_idx = first - 1;
+            ssize_t begin_idx = static_cast<ssize_t>(first) - 1;
 
             while (read_idx > begin_idx && !isNullAt(res[read_idx]))
             {
